+++ src/parser/error.rs
@@ -49,6 +49,7 @@ pub enum ParserError {
     UnexpectedReturn(usize, usize),             // (pos, line)
     InvalidContextForIdentifier(usize, String), // (line, value)
     InvalidTokenInFnCall(usize, usize, String), // (pos, line, value)
+    InvalidTokenInUnaryOp(usize, usize, String),// (pos, line, value)
 }
 
 impl Error for ParserError {}
@@ -123,6 +124,8 @@ impl fmt::Display for ParserError {
                 write!(f, "Parser error at source line {}: Expected ParseContext::FunctionCall or ParseContext::Normal, got {}", line, valuestr),
             ParserError::InvalidTokenInFnCall(pos, line, valuestr) =>
                 write!(f, "Parser error at source line {}, token position {}: Expected identifier or opening function argument bracket, instead of: '{}'", line, pos, valuestr),
+            ParserError::InvalidTokenInUnaryOp(pos, line, valuestr) =>
+                write!(f, "Parser error at source line {}, token position {}: Expected identifier, digit, boolean, or whitespace in unary operation, instead of: '{}'", line, pos, valuestr),
         }
     }
 }